Implement login exchange, email.py rename, main function
