+++ EMailService.py
@@ -1,6 +1,6 @@
 from abc import ABC, abstractmethod
 from dataclasses import dataclass
-from email import Email
+from email2 import Email
 
 
 class ProtocolTemplate(ABC):
@@ -53,15 +53,22 @@ class ImapProtocol(ProtocolTemplate):
 
 class ExchangeProtocol(ProtocolTemplate):
     
+    username : str = ""
+    password : str = ""
+
     @property
     def logged_in(self) -> bool:
-        pass
+        return self.username != "" and self.password != ""
 
     def login(self,user:str, password:str) -> bool:
-        pass
+        self.username = user
+        self.password = password
+        return True
     
     def logout(self) -> bool:
-        pass
+        self.username = ""
+        self.password = ""
+        return True
     
     def sendEmail(self,Email:Email) -> bool:
         """Requierment: User is logged in"""
@@ -73,3 +80,22 @@ class ExchangeProtocol(ProtocolTemplate):
     
     def getEmails(self)->list[Email]:
         pass
+
+if __name__ == "__main__":
+    imap = ImapProtocol()
+    exchange = ExchangeProtocol()
+
+    print("IMAP Logged_in: ",imap.logged_in)
+    imap.login("","")
+    print("IMAP Logged_in: ",imap.logged_in)
+    imap.logout()
+    print("IMAP Logged_in: ",imap.logged_in)
+
+    #exchange
+
+    print("Exchange Logged_in: ",exchange.logged_in)
+    exchange.login("ude-1729267167","6jTDTk6hS3j^b%@tw")
+    print("Exchange Logged_in: ",exchange.logged_in)
+    exchange.logout()
+    print("Exchange Logged_in: ",exchange.logged_in)
+
